requirements.txt
asgiref==3.8.1
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.2.0
colorama==0.4.6
Flask==3.1.1
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-MySQLdb==2.0.0
Flask-JWT-Extended==4.7.1 # ¡Actualizado a la versión 4.7.1!
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysqlclient==2.2.7
pillow==11.2.1
python-dotenv==1.1.0
redis==5.0.1
requests==2.32.3
smmap==5.0.2
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3